work on plot module; add test data; clean up readme
