+++ requirements-dev.txt
@@ -1,5 +1,5 @@
-pytest
-pytest-cov
-coverage
-flake8
-isort
+pytest==6.2.4
+pytest-cov==2.12.0
+coverage==5.5
+flake8==3.9.2
+isort==5.8.0
