+++ src/utils.py
@@ -1,10 +1,4 @@
-from cement.utils.version import get_version as cement_get_version
-
-VERSION = (0, 0, 1, 'alpha', 0)
-
-
-def get_version(version=VERSION):
-    return cement_get_version(version)
+VERSION = 'v0.1.0-alpha'
 
 
 class TastyworksCLIError(Exception):
